Make sandbox code compile as "chromium_code".
This sets up useful flags like -Wall -Werror etc.

Also squash a compiler warning:

sandbox/linux/suid/process_util_linux.c: In function ‘AdjustOOMScore’:
sandbox/linux/suid/process_util_linux.c:25: error: format ‘%lu’ expects type ‘long unsigned int’, but argument 4 has type ‘pid_t’

BUG=none
TEST=try-servers

Review URL: http://codereview.chromium.org/733001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@41161 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 012a67451afc724a41e26bd7998aa8523d294959

+++ sandbox/linux/suid/process_util_linux.c
@@ -8,6 +8,7 @@
 #include "process_util.h"
 
 #include <fcntl.h>
+#include <inttypes.h>
 #include <limits.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -20,9 +21,9 @@ bool AdjustOOMScore(pid_t process, int score) {
   if (score < 0 || score > 15)
     return false;
 
-  char oom_adj[35];  // "/proc/" + log_2(2**64) + "/oom_adj\0"
+  char oom_adj[35];  // "/proc/" + log_10(2**64) + "/oom_adj\0"
                      //    6     +       20     +     9         = 35
-  snprintf(oom_adj, sizeof(oom_adj), "/proc/%lu", process);
+  snprintf(oom_adj, sizeof(oom_adj), "/proc/%" PRIdMAX, (intmax_t)process);
 
   struct stat statbuf;
   if (stat(oom_adj, &statbuf) < 0)
